LexFlow_Integrated/requirements.txt
streamlit
openai
langchain
langchain-openai
langchain-community
langchain-text-splitters
chromadb
fastapi
uvicorn
requests
python-multipart
faiss-cpu
pydantic
python-docx
pdf2image
pytesseract
pdfplumber
xhtml2pdf
beautifulsoup4
google-search-results
python-dotenv
pandas
fastapi-mail
sentence-transformers
tqdm
duckduckgo-search
aiofiles
ddgs



pytesseract pillow pdf2image
tesseract poppler   # macOS
openai langchain-openai langchain-core